remodel component
